add comprar guitarrista e baixista

+++ Projeto_Java/src/br/edu/ufca/dados/Baterista.java
@@ -2,7 +2,8 @@ package br.edu.ufca.dados;
 
 public class Baterista extends MusicoAbstrato {
 	
-	public Baterista(String nome,int exp,int preco) {
+	public Baterista(String nome,double exp,int preco) {
 		super(nome,exp,preco);
 	}
+	
 }
