+++ Projeto_Java/src/br/edu/ufca/dados/Guitarrista.java
@@ -2,7 +2,7 @@ package br.edu.ufca.dados;
 
 public class Guitarrista extends MusicoAbstrato{
 	
-	public Guitarrista(String nome,int exp,int preco) {
+	public Guitarrista(String nome,double exp,int preco) {
 		super(nome,exp,preco);
 	}
 }
